requirements.txt
modal
newspaper3k
googlesearch-python
